changes done in sf-paymentmanagement

+++ backend/sf-paymentmanagement/src/main/java/com/pg/payment/dto/PaymentRequest.java
@@ -1,13 +1,14 @@
 package com.pg.payment.dto;
 
 import java.math.BigDecimal;
-import com.pg.payment.model.Payment.PaymentMethodEnum;
+
+import com.pg.payment.model.PaymentMethodEnum;
+
 import lombok.Data;
 
 @Data
 public class PaymentRequest {
-	private Integer bookingId;
+	private Long bookingId;
 	private BigDecimal amount;
 	private PaymentMethodEnum method;
-
 }
